# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_monitoring-system_backend_internal_domain_issue.CreateIssueRequest:
    properties:
      due_date:
        type: string
      issue_pic_user_id:
        type: string
      keterangan:
        type: string
      result_id:
        type: string
    required:
    - issue_pic_user_id
    - result_id
    type: object
  github_com_monitoring-system_backend_internal_domain_issue.IssueStatus:
    enum:
    - Open
    - InProgress
    - PendingValidation
    - Closed
    - Verified
    type: string
    x-enum-varnames:
    - IssueStatusOpen
    - IssueStatusInProgress
    - IssueStatusPendingValidation
    - IssueStatusClosed
    - IssueStatusVerified
  github_com_monitoring-system_backend_internal_domain_issue.UpdateIssueRequest:
    properties:
      due_date:
        type: string
      issue_pic_user_id:
        type: string
      issue_status:
        allOf:
        - $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_issue.IssueStatus'
        enum:
        - Open
        - InProgress
        - PendingValidation
        - Closed
        - Verified
      keterangan:
        type: string
    type: object
...
  github_com_monitoring-system_backend_pkg_response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  internal_handler_issuehandler.ExtendDueDateRequest:
    properties:
      due_date:
        type: string
    required:
    - due_date
    type: object
host: localhost:8080
info:
  contact:
    email: dev@company.com
    name: Internal Dev Team
  description: API untuk sistem monitoring audit internal perusahaan
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Monitoring Audit API
  version: "1.0"
paths:
  /api/v1/issues:
    get:
      consumes:
      - application/json
      description: Get a paginated list of issues, optionally filtered by status or
        pic_user_id
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by PIC User ID
        in: query
        name: pic_user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all issues
      tags:
      - issues
    post:
      consumes:
      - application/json
      description: Create a new issue with the provided details
      parameters:
      - description: Create Issue Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_issue.CreateIssueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new issue
      tags:
      - issues
  /api/v1/issues/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an issue by its ID
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete an issue
      tags:
      - issues
    get:
      consumes:
      - application/json
      description: Get a specific issue by its ID
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get issue by ID
      tags:
      - issues
    put:
      consumes:
      - application/json
      description: Update an existing issue by its ID
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Issue Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_monitoring-system_backend_internal_domain_issue.UpdateIssueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update an issue
      tags:
      - issues
  /api/v1/issues/{id}/extend-due-date:
    patch:
      consumes:
      - application/json
      description: Extend the due date of a specific issue
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      - description: Extend Due Date Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_issuehandler.ExtendDueDateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Extend issue due date
      tags:
      - issues
  /api/v1/issues/{id}/photos:
    get:
      consumes:
      - application/json
      description: Get all photos associated with a specific issue
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get photos by issue ID
      tags:
      - issue-photos
    post:
      consumes:
      - multipart/form-data
      description: Upload a photo for a specific issue
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: string
      - description: Photo file
        in: formData
        name: photo
        required: true
        type: file
      - description: Photo Type
        in: formData
        name: photo_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload an issue photo
      tags:
      - issue-photos
  /api/v1/issues/photos/{photo_id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific issue photo by its ID
      parameters:
      - description: Photo ID
        in: path
        name: photo_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete an issue photo
      tags:
      - issue-photos
  /api/v1/logs/activity:
    get:
      consumes:
      - application/json
      description: Get a paginated list of activity logs, optionally filtered by user_id,
        module_id, or action
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by User ID
        in: query
        name: user_id
        type: string
      - description: Filter by Module ID
        in: query
        name: module_id
        type: string
      - description: Filter by Action
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all activity logs
      tags:
      - activity-logs
  /api/v1/logs/activity/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific activity log by its ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get activity log by ID
      tags:
      - activity-logs
  /api/v1/logs/login:
    get:
      consumes:
      - application/json
      description: Get a paginated list of login logs, optionally filtered by user_id
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by User ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all login logs
      tags:
      - login-logs
  /api/v1/logs/login/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific login log by its ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get login log by ID
      tags:
      - login-logs
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"